Iris/requirements.txt
streamlit==0.61.0
pandas==0.25.3
numpy==1.18.1
scikit-learn==0.22.1
pillow==8.1.1